### 2024-11-04 #7426 @editorialbot
[PRE REVIEW]: TimeStruct.jl - flexible multi-horizon time modelling in optimization models
**Submitting author:** @trulsf (<a href="http://orcid.org/0000-0002-5914-6122">Truls Flatberg</a>)
**Repository:** https://github.com/sintefore/TimeStruct.jl
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** 0.9.1
