### 2024-11-22 e77bdba
Add rowtables to test AOG
