### 2024-11-22 7603b67
Define TimeProfile before use
