### 2024-12-10 #7426 @vissarion
Hi, @vankesteren, @ChrisRackauckas, @tmigot, @ziyiyin97, @lucaferranti, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
