### 2024-12-10 #7426 @vissarion
Hi, @odow, @datejada, you were recommended by the submitting author as suitable reviewers for this submission.
Would you be willing to review this submission for JOSS?
