### 2024-12-11 #7578 @editorialbot
[REVIEW]: TimeStruct.jl - flexible multi-horizon time modelling in optimization models
**Submitting author:** @trulsf (<a href="http://orcid.org/0000-0002-5914-6122">Truls Flatberg</a>)
**Repository:** https://github.com/sintefore/TimeStruct.jl
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v0.9.2
