### 2025-01-20 #7578 @datejada
@trulsf I have reviewed the checklist, and overall, it is a great and helpful package to complement JuMP functionalities. Thanks!

A couple of minor things from my side:

- The example in the readme is not working. I have created a PR to fix it: https://github.com/sintefore/TimeStruct.jl/pull/49
- The documentation is missing the statement of need; maybe you can add a section with the information that is already in the paper
- The paper is missing the "state of the field" section on the checklist; please include it.

Apart from this, I don't have more comments.

CC: @vissarion
