### 2025-01-27 sintefore/TimeStruct.jl#50 @trulsf
JOSS review
Based on feedback from the first review (https://github.com/openjournals/joss-reviews/issues/7578#issuecomment-2601860011), the following issues should be addressed:
- [x] Fix example in README
- [x] Review state of the field in paper
- [x] Add statement of need to pacakage documentation
- [x] Fix example formatting in paper
