### 2025-01-27 4f15a0c
Add short state of the field to joss paper and fix format
