### 2025-01-27 4068d0f
Add explanation for TimeStruct benefits in documentation
