### 2025-01-27 6d7b40a
Apply suggestions from code review
Co-authored-by: hellemo <lars.hellemo@sintef.no>
Co-authored-by: Julian Straus <104911227+JulStraus@users.noreply.github.com>
