+++ docs/papers/joss/paper.bib
@@ -164,6 +164,7 @@ DOI = {10.3390/en10101468}
   number={3},
   pages={729--746},
   year={2016},
+  doi={10.1016/j.ejor.2016.06.039},
   publisher={Elsevier}
 }
 
@@ -175,5 +176,6 @@ DOI = {10.3390/en10101468}
   number={4},
   pages={1380--1397},
   year={1998},
+  doi={10.1021/ie970532x},
   publisher={ACS Publications}
 }
