### 2025-01-27 sintefore/TimeStruct.jl#50 @trulsf
This issue is fixed by #49 #51 and #52.
