### 2025-01-27 8ba3f38
Add short state of the field to joss paper and fix format (#51)
* Add short state of the field to joss paper and fix format

* Apply suggestions from code review

Co-authored-by: hellemo <lars.hellemo@sintef.no>
Co-authored-by: Julian Straus <104911227+JulStraus@users.noreply.github.com>

---------

Co-authored-by: hellemo <lars.hellemo@sintef.no>
Co-authored-by: Julian Straus <104911227+JulStraus@users.noreply.github.com>
