+++ docs/papers/joss/paper.bib
@@ -155,3 +155,27 @@ DOI = {10.3390/en10101468}
   year={2014},
   publisher={Springer}
 }
+
+@article{correia2016multi,
+  title={Multi-period capacitated facility location under delayed demand satisfaction},
+  author={Correia, Isabel and Melo, Teresa},
+  journal={European Journal of Operational Research},
+  volume={255},
+  number={3},
+  pages={729--746},
+  year={2016},
+  doi={10.1016/j.ejor.2016.06.039},
+  publisher={Elsevier}
+}
+
+@article{iyer1998optimal,
+  title={Optimal planning and scheduling of offshore oil field infrastructure investment and operations},
+  author={Iyer, RR and Grossmann, IE and Vasantharajan, S and Cullick, AS},
+  journal={Industrial \& Engineering Chemistry Research},
+  volume={37},
+  number={4},
+  pages={1380--1397},
+  year={1998},
+  doi={10.1021/ie970532x},
+  publisher={ACS Publications}
+}
