### 2025-01-27 #7578 @trulsf
@datejada Thanks for your review and positive feedback! We have updated the package and the paper based upon your comments. Thanks for your PR which is now merged. For further details on the updates see https://github.com/sintefore/TimeStruct.jl/pull/51 https://github.com/sintefore/TimeStruct.jl/pull/52
