### 2025-01-27 ac6a3e6
fix some typos in the docs
