### 2025-01-27 a4d89a1
fix some typos in the docs
