### 2025-01-27 #7578 @lucaferranti
Hi @trulsf  (cc @vissarion )

First of all, apologies for the delay with the review.

Overall, I think the repository quality is very high and the paper is well written and as such I recommend publication.

### Improvement suggestions

- Currently, the examples in the documentation are relatively "low-level" in the sense that they only show how to build the optimization. I think it would be very good to have a bigger "end-to-end" application example, which shows how `TimeStruct.jl` is used to solve an optimization problem.  Since the paper briefly mentions some example applications in the last section without further details, it would be nice to have the code for those examples in the documentation and link there from the paper.
- I opened a PR fixing some typos I found while going through the documentation
- (minor nitpick) If I'm not mistaken, looking at the CI job, it seems the repository uses codecov, why not adding the coverage badge?
- The documentation does not have a section called "Statement of need", but the "Why to use TimeStruct" section does convey the same message clearly and I don't see a renaming necessarily, but I'll let the editor have the last word on this.

- The paper is slightly inconsistent in how it uses Julia packages names. Throughout most of the paper it uses `TimeStruct` for the submitted software (normal font and no `.jl`), in line 57 it is `TimeStruct` with different font and in line 67 it uses `TimeStruct.jl` (with the `.jl` and different font, I believe obtained by using backticks in markdown).  `EnergyModelsBase.jl` is named also with different font. I would recommend to homogenize these through the text, particularly I would recommend the format of line 67, i.e. include the `.jl` in the name and use backticks.

Other than these minor observations, i think it's a great work and recommend for publication!
