### 2025-01-28 #7578 @datejada
> [@datejada](https://github.com/datejada) Thanks for your review and positive feedback! We have updated the package and the paper based upon your comments. Thanks for your PR which is now merged. For further details on the updates see [sintefore/TimeStruct.jl#51](https://github.com/sintefore/TimeStruct.jl/pull/51) [sintefore/TimeStruct.jl#52](https://github.com/sintefore/TimeStruct.jl/pull/52)

@trulsf, thanks for including the suggestions; I have no more comments. @vissarion, I have checked all the boxes on the list. Please let me know if I need to do anything else. Thanks! Diego
