### 2025-01-28 accbaf5
Fixed a bug for multiple_strat
