+++ test/runtests.jl
@@ -581,6 +581,8 @@ end
     per = ops2[1]
     @test typeof(per) <: TimeStruct.SimplePeriod
 
+    @test multiple_strat(first(strat_periods(simple)), first(ops1)) == 1/10
+
     ops3 = [t for sc in opscenarios(simple) for t in sc]
     @test ops1 == ops3
 end
