### 2025-01-28 6560aec
Fixed a bug for multiple_strat (#54)
