### 2025-01-29 sintefore/TimeStruct.jl#55 @trulsf
JOSS review 2
Based on feedback from the second review (https://github.com/openjournals/joss-reviews/issues/7578#issuecomment-2616831519), the following issues should be addressed:
- [x] Improved application example for an optimization problem in the docs and links in the paper
- [x] Add coverage badge
- [x] Merge PR with doc fixes
- [x] Consistent formatting of package names in paper
