### 2025-01-29 #7578 @trulsf
@lucaferranti Thanks for the good feedback and help with a PR! We have opened an issue (https://github.com/sintefore/TimeStruct.jl/issues/55) to address the comments raised by the review and will inform you when we are finished.
