+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"manual/profiles/#man-prof","page":"Time profiles","title":"Time profiles","text":"","category":"section"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"To provide data for different time structures there is a flexible system of different time profiles that can be indexed by time periods.","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"FixedProfile: Time profile with the same value for all time periods\nOperationalProfile: Time profile with values varying with operational time periods\nScenarioProfile: Holds a separate time profile for each operational scenario\nRepresentativeProfile: Holds a separate time profile for each representative period\nStrategicProfile : Holds a separate time profile for each strategic period\nStrategicStochasticProfile : Holds a separate time profile for each strategic node in a strategic tree","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"The following code example shows how these profile types can be combined in a flexible manner to produce different overall profiles.","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"rep_periods = RepresentativePeriods(2, 365, [0.6, 0.4], [SimpleTimes(7,1), SimpleTimes(7,1)])\nperiods = TwoLevel(2, 365, rep_periods)\n\ncost = StrategicProfile(\n            [\n                RepresentativeProfile(\n                    [\n                        OperationalProfile([3, 3, 4, 3, 4, 6, 5]),\n                        FixedProfile(5)\n                    ]\n                ),\n                FixedProfile(7)\n            ]\n        )","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"Illustration of profile values for the various time periods as defined in the profile example (Image: Time profile values)","category":"page"},{"location":"manual/discount/#man-disc","page":"Discounting","title":"Discounting","text":"","category":"section"},{"location":"manual/discount/","page":"Discounting","title":"Discounting","text":"For multi-year investment optimization models it is common pratice to use an objective that is discounted to get the net present value of the investment. Since investment decisions usually are done on a strategic level, discount factors are also calculated based on strategic periods.","category":"page"},{"location":"manual/discount/","page":"Discounting","title":"Discounting","text":"The discount factor for a time period t is found by the discount function. There are two strategies for calculating the discount factor, either all discounting is calculated based on the start of the strategic period or it is based on finding an approximation of the average value over the strategic period. The following example shows how these two types will differ for a planning period of 50 years, consisting of 5 periods of 10 years:","category":"page"},{"location":"manual/discount/","page":"Discounting","title":"Discounting","text":"using TimeStruct\nts = TwoLevel(5, 10, SimpleTimes(1,1));\ndf_start = [discount(t, ts, 0.05; type = \"start\") for t in ts]\ndf_avg = [discount(t, ts, 0.05; type = \"avg\") for t in ts]","category":"page"},{"location":"manual/discount/","page":"Discounting","title":"Discounting","text":"While it is often normal to assume investments at the start of each strategic period, it can be more correct to average the discount factor for operational costs that are accrued throughout the strategic period.","category":"page"},{"location":"manual/discount/","page":"Discounting","title":"Discounting","text":"To help setting up the objective function in a typical optimization problem, there is a utility function objective_weight that returns the weight to give a time period in the objective, considering both discount factor, probability and possible multiplicity.","categor
[…]
