### 2025-01-29 sintefore/TimeStruct.jl#55 @JulStraus
Will take a look at the package names in the paper.
