### 2025-02-02 f344b24
Add tutorials for battery sizing and EnergyModelsX; include Literate  support in docs
