+++ docs/src/tutorials/battery_sizing.jl
@@ -1,0 +1,211 @@
+# # Battery sizing
+
+
+# This tutorial demonstrates how to formulate a battery sizing problem using the `TimeStruct` package
+# in combination with `JuMP`.
+# We start by defining the operational model of the battery, which includes the constraints and the
+# objective function. Then, we extend the model to include the strategic decisions of the battery sizing.
+
+
+# For this tutorial, we use the `HiGHS` solver for optimization:
+using TimeStruct
+using JuMP
+using HiGHS
+
+# ## Operational model
+
+# We start by defining the operational model of the battery.
+# The battery has a fixed capacity and can charge and discharge energy to cover
+# a given demand profile. Additionally, the demand can be covered by purchasing
+# energy from the spot market with a given price profile.
+function create_operational(periods::TimeStructure, capacity, demand::TimeProfile, price::TimeProfile)
+    model = Model()
+
+    @variable(model, soc[periods] >= 0)         # State of charge of the battery
+    @variable(model, charge[periods] >= 0)      # Energy charged to the battery
+    @variable(model, discharge[periods] >= 0)   # Energy discharged from the battery
+    @variable(model, spot[periods] >= 0)        # Energy purchased from the spot market
+
+    for (prev, t) in withprev(periods)
+        @constraint(model, soc[t] <= capacity)
+        soc_prev = isnothing(prev) ? soc[last(periods)] : soc[prev]
+        @constraint(model, soc[t] == soc_prev + charge[t] - discharge[t])
+        @constraint(model, spot[t] + discharge[t] - charge[t] == demand[t])
+    end
+
+    @objective(model, Min, sum(multiple(t) * price[t] * spot[t] for t in periods))
+    return model
+end
+# Note the use of the special iterator  [`withprev`](@ref) to iterate over the periods with the previous period,
+# returning `nothing` for the first period. In this example we have opted to use a cyclic constraint for the state of charge,
+# i.e. the state of charge of the last period is used as the initial state of charge for the first period.
+
+# To test the model, we create a simple example with 24 hourly periods, a capacity of 20, a fixed demand of 5.0,
+# and a time dependent price profile that varies sinusoidally.
+periods = SimpleTimes(24, 1)
+capacity = 20
+demand = FixedProfile(5.0)
+price = OperationalProfile([1 + 0.3 * sin(i) for (i,t) in enumerate(periods)])
+
+model = create_operational(periods, capacity, demand, price)
+set_optimizer(model, HiGHS.Optimizer)
+set_silent(model)
+optimize!(model)
+println("Objective value: ", objective_value(model))
+
+
+
+# ## Strategic model
+
+# We now extend the operational model to include the strategic decisions of the battery sizing.
+# For simplicity, we assume that the battery capacity can be chosen independently for each period,
+# e.g. by renting a battery of a given capacity for each period.
+
+
+# To simplify the process of defining the strategic model, we define helper functions to create the operational variables,
+# constraints, and objective function.
+function create_operational_variables(model, periods)
+    @variable(model, soc[periods] >= 0)
+    @variable(model, charge[periods] >= 0)
+    @variable(model, discharge[periods] >= 0)
+    @variable(model, spot[periods] >= 0)
+ end
+
+function create_operational_constraints(model, periods, capacity, demand)
+    soc, charge, discharge = model[:soc], model[:charge], model[:discharge]
+    spot = model[:spot]
+    for (prev, t) in withprev(periods)
+        @constraint(model, soc[t] <= capacity)
[…]
