+++ docs/src/tutorials/emx.md
@@ -1,0 +1,1 @@
+# EnergyModelsX
