+++ docs/src/tutorials/battery_sizing.jl
@@ -1,11 +1,9 @@
 # # Battery sizing
 
-
 # This tutorial demonstrates how to formulate a battery sizing problem using the `TimeStruct` package
 # in combination with `JuMP`.
 # We start by defining the operational model of the battery, which includes the constraints and the
 # objective function. Then, we extend the model to include the strategic decisions of the battery sizing.
-
 
 # For this tutorial, we use the `HiGHS` solver for optimization:
 using TimeStruct
@@ -18,7 +16,12 @@ using HiGHS
 # The battery has a fixed capacity and can charge and discharge energy to cover
 # a given demand profile. Additionally, the demand can be covered by purchasing
 # energy from the spot market with a given price profile.
-function create_operational(periods::TimeStructure, capacity, demand::TimeProfile, price::TimeProfile)
+function create_operational(
+    periods::TimeStructure,
+    capacity,
+    demand::TimeProfile,
+    price::TimeProfile,
+)
     model = Model()
 
     @variable(model, soc[periods] >= 0)         # State of charge of the battery
@@ -45,7 +48,7 @@ end
 periods = SimpleTimes(24, 1)
 capacity = 20
 demand = FixedProfile(5.0)
-price = OperationalProfile([1 + 0.3 * sin(i) for (i,t) in enumerate(periods)])
+price = OperationalProfile([1 + 0.3 * sin(i) for (i, t) in enumerate(periods)])
 
 model = create_operational(periods, capacity, demand, price)
 set_optimizer(model, HiGHS.Optimizer)
@@ -53,14 +56,11 @@ set_silent(model)
 optimize!(model)
 println("Objective value: ", objective_value(model))
 
-
-
 # ## Strategic model
 
 # We now extend the operational model to include the strategic decisions of the battery sizing.
 # For simplicity, we assume that the battery capacity can be chosen independently for each period,
 # e.g. by renting a battery of a given capacity for each period.
-
 
 # To simplify the process of defining the strategic model, we define helper functions to create the operational variables,
 # constraints, and objective function.
@@ -69,7 +69,7 @@ function create_operational_variables(model, periods)
     @variable(model, charge[periods] >= 0)
     @variable(model, discharge[periods] >= 0)
     @variable(model, spot[periods] >= 0)
- end
+end
 
 function create_operational_constraints(model, periods, capacity, demand)
     soc, charge, discharge = model[:soc], model[:charge], model[:discharge]
@@ -91,7 +91,12 @@ end
 # structures where the operational periods do not cover the complete strategic periods.
 
 # We can now define the strategic model by creating a battery sizing decision for each strategic period.
-function create_strategic_model(periods::TimeStructure, price::TimeProfile, demand::TimeProfile, capex)
+function create_strategic_model(
+    periods::TimeStructure,
+    price::TimeProfile,
+    demand::TimeProfile,
+    capex,
+)
     model = Model()
     @variable(model, cap[strat_periods(periods)] >= 0)
 
@@ -109,9 +114,9 @@ end
 
 # To test the model, we create a simple example with 7 strategic periods of 24 hourly periods each
 periods = TwoLevel(7, SimpleTimes(24, 1))
-demand = StrategicProfile([5.0 + 0.5 * i for (i,_) in enumerate(strat_periods(periods))])
[…]
