### 2025-02-03 9471d8f
Add HiGHS dependency to Project.toml
