+++ docs/src/tutorials/emx.md
@@ -1,1 +1,84 @@
-# EnergyModelsX
+# [EnergyModelsX](@id tut-emx)
+
+**[EnergyModelsX (EMX)](https://github.com/EnergyModelsX)** is a flexible multi-horizon energy system optimization framework using [`JuMP`](https://jump.dev/JuMP.jl/stable/).
+It utilizes `TimeStruct` and its precursor `TimeStructures` from the beginning although some functionality from `TimeStruct` was not available in the initial version ([`OperationalScenarios`](@ref man-oper-osc), [`RepresentativePeriods`](@ref man-oper-repr), and [`TwoLevelTree](@ref man-multi-twoleveltree)).
+However, due to the structure of `TimeStruct`, it was simple to implement the functionality for the different time structures.
+This is outlined in the following sections.
+
+## [Implementation of the core structures](@id tut-emx-core)
+
+The core package `EnergyModelsBase` differentiates between variables indexed over strategic periods and variables indexed over operational period, similar to the battery sizing example.
+As an example, consider the following two functions for variable declaration for a given `𝒯::TimeStructure`:
+
+- [`variables_capacity(m, 𝒩::Vector{<:Node}, 𝒳ᵛᵉᶜ, 𝒯, modeltype::EnergyModel)`](https://github.com/EnergyModelsX/EnergyModelsBase.jl/blob/10036d91949e69cae0fbdb8e1652b85d7f82742a/src/model.jl#L175) with, among others,
+
+  ```julia
+  @variable(m, cap_use[𝒩ᶜᵃᵖ, 𝒯] >= 0)
+  @variable(m, cap_inst[𝒩ᶜᵃᵖ, 𝒯] >= 0)
+  ```
+
+- [`variables_opex(m, 𝒩::Vector{<:Node}, 𝒳ᵛᵉᶜ, 𝒯, modeltype::EnergyModel)`](https://github.com/EnergyModelsX/EnergyModelsBase.jl/blob/10036d91949e69cae0fbdb8e1652b85d7f82742a/src/model.jl#L297C10-L297C82) with, among others,
+
+  ```julia
+  𝒯ᴵⁿᵛ = strategic_periods(𝒯)
+  @variable(m, opex_var[𝒩ᵒᵖᵉˣ, 𝒯ᴵⁿᵛ])
+  ```
+
+These two functions highlight the simplicity of using `TimeStruct` as the individual time structures allow for iterations within `JuMP` macros.
+In addition, it simplifies the index sets as changing the number of intermediate time structures through incorporating, *e.g.*, [`OperationalScenarios`](@ref man-oper-osc), does not require changes to the variable declarations.
+The change in the structure through the incorporation of the operational scenarios is instead available within the individual `TimePeriod` type.
+
+The same is also true for constraint functions as shown in [`constraints_capacity(m, n::Node, 𝒯::TimeStructure, modeltype::EnergyModel)`](https://github.com/EnergyModelsX/EnergyModelsBase.jl/blob/10036d91949e69cae0fbdb8e1652b85d7f82742a/src/constraint_functions.jl#L17):
+
+```julia
+@constraint(m, [t ∈ 𝒯], m[:cap_use][n, t] <= m[:cap_inst][n, t])
+```
+
+Variables whose values are dependent on the previous operational period are only available for `Storage` nodes.
+Its implementation is rather complex, but it is entirely relying on the [`withprev`](@ref man-iter-prev) functionality to decide whether it is the first operational period in a different `TimeStructure` as well as how it must behave in this situation.
+
[…]
