### 2025-02-03 dedddfb
Add codecov badge with current setup.
