+++ docs/make.jl
@@ -14,6 +14,7 @@ end
 pages = [
     "Introduction" => "index.md",
     "Tutorials" => [
+        "Basic usage" => "tutorials/basic_usage.md",
         "Battery sizing" => "tutorials/battery_sizing.md",
         "EnergyModelsX" => "tutorials/emx.md",
     ],
