### 2025-02-04 eb7a931
Extend battery sizing tutorial
