+++ docs/src/tutorials/battery_sizing.jl
@@ -10,6 +10,8 @@ using TimeStruct
 using JuMP
 using HiGHS
 
+optimizer = optimizer_with_attributes(HiGHS.Optimizer, "output_flag" => false)
+
 # ## Operational model
 
 # We start by defining the operational model of the battery.
@@ -44,23 +46,22 @@ end
 # i.e. the state of charge of the last period is used as the initial state of charge for the first period.
 
 # To test the model, we create a simple example with 24 hourly periods, a capacity of 20, a fixed demand of 5.0,
-# and a time dependent price profile that varies sinusoidally.
+# and a time dependent price profile that varies sinusoidally over a day.
 periods = SimpleTimes(24, 1)
 capacity = 20
 demand = FixedProfile(5.0)
 price = OperationalProfile([1 + 0.3 * sin(i) for (i, t) in enumerate(periods)])
 
 model = create_operational(periods, capacity, demand, price)
-set_optimizer(model, HiGHS.Optimizer)
-set_silent(model)
+set_optimizer(model, optimizer)
 optimize!(model)
 println("Objective value: ", objective_value(model))
 
 # ## Strategic model
 
 # We now extend the operational model to include the strategic decisions of the battery sizing.
-# For simplicity, we assume that the battery capacity can be chosen independently for each period,
-# e.g. by renting a battery of a given capacity for each period.
+# For simplicity, we assume that the battery capacity can be chosen independently for each stragic period,
+# e.g., by renting a battery of a given capacity for each strategic period.
 
 # To simplify the process of defining the strategic model, we define helper functions to create the operational variables,
 # constraints, and objective function.
@@ -84,13 +85,15 @@ end
 
 function create_operational_objective(model, periods, price)
     spot = model[:spot]
-    el_cost = sum(multiple(t) * price[t] * spot[t] for t in periods)
+    el_cost = sum(multiple(t) * probability(t) * price[t] * spot[t] for t in periods)
     return el_cost
 end
 # Note the use of the [`multiple`](@ref) function to include the multiplier of each operational period for time
-# structures where the operational periods do not cover the complete strategic periods.
+# structures where the operational periods do not cover the complete strategic periods and
+# [`probability`](@ref) for periods that have an associated probability when using operational scenarios.
 
 # We can now define the strategic model by creating a battery sizing decision for each strategic period.
+# To obtain the strategic periods, we use the [`strat_periods`](@ref) iterator.
 function create_strategic_model(
     periods::TimeStructure,
     price::TimeProfile,
@@ -114,22 +117,23 @@ end
 
 # To test the model, we create a simple example with 7 strategic periods of 24 hourly periods each
 periods = TwoLevel(7, SimpleTimes(24, 1))
-demand = StrategicProfile([5.0 + 0.5 * i for (i, _) in enumerate(strat_periods(periods))])
-price = OperationalProfile([1 + 0.3 * sin(i) for (i, t) in enumerate(periods)])
-capex = StrategicProfile([2 - 0.1 * i for (i, _) in enumerate(strat_periods(periods))])
+demand = StrategicProfile([5.0 + 0.5 * i for i in 1:7])
+price = OperationalProfile([1 + 0.3 * sin(i) for i in 1:24])
+capex = StrategicProfile([2 - 0.1 * i for i in 1:7])
 
 model = create_strategic_model(periods, price, demand, capex)
 
-set_optimizer(model, HiGHS.Optimizer)
-set_silent(model)
+set_optimizer(model, optimizer)
 optimize!(model)
 println("Objective value: ", objective_value(model))
-println("Battery sizing: ", value.(model[:cap]))
+println("Battery sizing: ", [value(model[:cap][sp]) for sp in strat_periods(periods)])
 
-# ### RepresentativePeriods
+# ### Representative periods
 
[…]
