+++ docs/make.jl
@@ -15,7 +15,9 @@ pages = [
     "Introduction" => "index.md",
     "Tutorials" => [
         "Battery sizing" => "tutorials/battery_sizing.md",
-        "EnergyModelsX" => "tutorials/emx.md",
+    ],
+    "Showcases" => [
+        "EnergyModelsX" => "showcases/emx.md",
     ],
     "Manual" => [
         "Operational time structures" => "manual/basic.md",
