+++ docs/src/showcases/emx.md
@@ -1,11 +1,10 @@
-# [EnergyModelsX](@id tut-emx)
+# [EnergyModelsX](@id show-emx)
 
 **[EnergyModelsX (EMX)](https://github.com/EnergyModelsX)** is a flexible multi-horizon energy system optimization framework using [`JuMP`](https://jump.dev/JuMP.jl/stable/).
-It utilizes `TimeStruct` and its precursor `TimeStructures` from the beginning although some functionality from `TimeStruct` was not available in the initial version ([`OperationalScenarios`](@ref man-oper-osc), [`RepresentativePeriods`](@ref man-oper-repr), and [`TwoLevelTree](@ref man-multi-twoleveltree)).
-However, due to the structure of `TimeStruct`, it was simple to implement the functionality for the different time structures.
-This is outlined in the following sections.
+It utilizes `TimeStruct` from the beginning and influenced some of the added features of `TimeStruct`.
+Its implemtation is outlined in the following sections.
 
-## [Implementation of the core structures](@id tut-emx-core)
+## [Implementation of the core structures](@id show-emx-core)
 
 The core package `EnergyModelsBase` differentiates between variables indexed over strategic periods and variables indexed over operational period, similar to the battery sizing example.
 As an example, consider the following two functions for variable declaration for a given `𝒯::TimeStructure`:
@@ -45,9 +44,9 @@ duration(t) * multiple\_strat(t_{inv}, t) * probability(t)
 
 where ``t`` corresponds to an operational period and ``t_{inv}`` to a strategic period.
 
-## [Advanced utilization](@id tut-emx-adv)
+## [Advanced utilization](@id show-emx-adv)
 
-### [`withprev` and its application](@id tut-emx-adv-withprev)
+### [`withprev` and its application](@id show-emx-adv-withprev)
 
 As outlined, neither [`OperationalScenarios`](@ref man-oper-osc) or [`RepresentativePeriods`](@ref man-oper-repr) where available in the initial development of **EMX**.
 However, the adjustments towards including these were limited to constraints that require the previous periods, that is constraints declared through the [`withprev`](@ref man-iter-prev) functionality.
@@ -66,19 +65,19 @@ Due to the iteration (and storing) of all previous periods, it is possible to id
 
     If you do not have any constraints depending on the previous periods, you even do not have to do any changes to your model when including `OperationalScenarios` or `RepresentativePeriods`.
 
-### [Chunks for unit commitment](@id tut-emx-adv-chunk)
+### [Chunks for unit commitment](@id show-emx-adv-chunk)
 
-The functionality of `TimeStruct` for chuncks based on the minimum required time is utilized in the *[unit commitment constraints of `Reformer` nodes](https://github.com/EnergyModelsX/EnergyModelsHydrogen.jl/blob/88a21ffae88b7ce199752aa5465313ad549718b6/src/constraints/reformer.jl#L259)* in [`EnergyModelsHydrogen`](https://energymodelsx.github.io/EnergyModelsHydrogen.jl/stable/).
+The functionality of `TimeStruct` for chunks based on the minimum required time is utilized in the *[unit commitment constraints of `Reformer` nodes](https://github.com/EnergyModelsX/EnergyModelsHydrogen.jl/blob/88a21ffae88b7ce199752aa5465313ad549718b6/src/constraints/reformer.jl#L259)* in [`EnergyModelsHydrogen`](https://energymodelsx.github.io/EnergyModelsHydrogen.jl/stable/).
 In this context, we require a minimum time for starting the node, shutting the node down, and when the node is offline due to limitations in the dynamics of the chemical plant.
 
 The used approach is similar to the `Storage` level balance utilizing the [`withprev`](@ref man-iter-prev) functionality for the majority of the time structures.
[…]
