+++ docs/src/tutorials/emx.md
@@ -1,1 +1,0 @@
-# EnergyModelsX
