### 2025-02-05 a046ca1
Adding tutorials (#57)
* Add tutorials for basic usage and battery sizing 

* Include Literate.jl  support in docs

* Added the EMX tutorial (#58) to showcase

---------

Co-authored-by: Julian Straus <104911227+JulStraus@users.noreply.github.com>
