+++ docs/make.jl
@@ -1,7 +1,23 @@
-using Documenter, TimeStruct
+using Documenter
+using Literate
+using TimeStruct
+
+# Literate on tutorial folder
+for (root, dir, files) in walkdir(joinpath(@__DIR__, "src", "tutorials"))
+    for file in files
+        if endswith(file, ".jl")
+            Literate.markdown(joinpath(root, file), root; documenter = true)
+        end
+    end
+end
 
 pages = [
     "Introduction" => "index.md",
+    "Tutorials" => [
+        "Basic usage" => "tutorials/basic_usage.md",
+        "Battery sizing" => "tutorials/battery_sizing.md",
+    ],
+    "Showcases" => ["EnergyModelsX" => "showcases/emx.md"],
     "Manual" => [
         "Operational time structures" => "manual/basic.md",
         "Multi-horizon" => "manual/multi.md",
