### 2025-02-05 sintefore/TimeStruct.jl#55 @trulsf
Closed through the work in #53 #56 #57 #58 #59
