### 2025-02-05 #7578 @trulsf
@lucaferranti We have now revised the paper and documentation based on your feedback (see https://github.com/sintefore/TimeStruct.jl/issues/55). The documentation is extended with two tutorials (one on basic usage and one on battery sizing) and a showcase section that describes how TimeStruct is used in EnergyModelsX. These should help users by providing "end-to-end" examples on how TimeStruct is used to create optimization models. Our plan is to extend with more tutorials and showcases going forward.  A coverage badge is now added and we have revised the paper to be consistent in the way package names are used. Once again, thanks for the PR!

@vissarion We leave it up to you if there is a need to change the section header to "Statement of need".
