### 2025-02-05 #7578 @lucaferranti
Thank you @trulsf for addressing my comments. The new tutorials are very nice and I think they'll help new users appreciate the package better :) I also like that you used Literate.jl for them, as it makes it easier for users to grab the code and play with it.

@vissarion this is ready from my side, all boxes are ticked
