+++ v0.9.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"manual/profiles/#man-prof","page":"Time profiles","title":"Time profiles","text":"","category":"section"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"To provide data for different time structures there is a flexible system of different time profiles that can be indexed by time periods.","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"FixedProfile: Time profile with the same value for all time periods\nOperationalProfile: Time profile with values varying with operational time periods\nScenarioProfile: Holds a separate time profile for each operational scenario\nRepresentativeProfile: Holds a separate time profile for each representative period\nStrategicProfile : Holds a separate time profile for each strategic period\nStrategicStochasticProfile : Holds a separate time profile for each strategic node in a strategic tree","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"The following code example shows how these profile types can be combined in a flexible manner to produce different overall profiles.","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"rep_periods = RepresentativePeriods(2, 365, [0.6, 0.4], [SimpleTimes(7,1), SimpleTimes(7,1)])\nperiods = TwoLevel(2, 365, rep_periods)\n\ncost = StrategicProfile(\n            [\n                RepresentativeProfile(\n                    [\n                        OperationalProfile([3, 3, 4, 3, 4, 6, 5]),\n                        FixedProfile(5)\n                    ]\n                ),\n                FixedProfile(7)\n            ]\n        )","category":"page"},{"location":"manual/profiles/","page":"Time profiles","title":"Time profiles","text":"Illustration of profile values for the various time periods as defined in the profile example (Image: Time profile values)","category":"page"},{"location":"showcases/emx/#show-emx","page":"EnergyModelsX","title":"EnergyModelsX","text":"","category":"section"},{"location":"showcases/emx/","page":"EnergyModelsX","title":"EnergyModelsX","text":"EnergyModelsX (EMX) is a flexible multi-horizon energy system optimization framework using JuMP. It utilizes TimeStruct from the beginning and influenced some of the added features of TimeStruct. Its implemtation is outlined in the following sections.","category":"page"},{"location":"showcases/emx/#show-emx-core","page":"EnergyModelsX","title":"Implementation of the core structures","text":"","category":"section"},{"location":"showcases/emx/","page":"EnergyModelsX","title":"EnergyModelsX","text":"The core package EnergyModelsBase differentiates between variables indexed over strategic periods and variables indexed over operational period, similar to the battery sizing example. As an example, consider the following two functions for variable declaration for a given 𝒯::TimeStructure:","category":"page"},{"location":"showcases/emx/","page":"EnergyModelsX","title":"EnergyModelsX","text":"variables_capacity(m, 𝒩::Vector{<:Node}, 𝒳ᵛᵉᶜ, 𝒯, modeltype::EnergyModel) with, among others,\n@variable(m, cap_use[𝒩ᶜᵃᵖ, 𝒯] >= 0)\n@variable(m, cap_inst[𝒩ᶜᵃᵖ, 𝒯] >= 0)\nvariables_opex(m, 𝒩::Vector{<:Node}, 𝒳ᵛᵉᶜ, 𝒯, modeltype::EnergyModel) with, among others,\n𝒯ᴵⁿᵛ = strategic_periods(𝒯)\n@variable(m, opex_var[𝒩ᵒᵖᵉˣ, 𝒯ᴵⁿᵛ])","category":"page"},{"location":"showcases/emx/","page":"EnergyModelsX","title":"EnergyModelsX","text":"These two 
[…]
