### 2025-02-13 #7578 @trulsf
@vissarion Thanks! We have tagged a new version (v0.9.2) with the changes from the review. The newest release is archived at zenodo at https://doi.org/10.5281/zenodo.14867823
