### 2025-02-14 #7578 @trulsf
I have updated the order and affiliations in zenodo. Hopefully, the information is now aligned with the paper.
