### 2025-02-14 dfea22c
Remove duplicated entry from bibliography
