+++ docs/papers/joss/paper.bib
@@ -71,16 +71,6 @@
     doi = {10.1007/s12532-023-00239-3}
 }
 
-@article{bodal2024hydrogen,
-  title={Hydrogen for harvesting the potential of offshore wind: A North Sea case study},
-  author={B{\o}dal, Espen Flo and Holm, Sigmund Eggen and Subramanian, Avinash and Durakovic, Goran and Pinel, Dimitri and Hellemo, Lars and Ortiz, Miguel Mu{\~n}oz and Knudsen, Brage Rugstad and Straus, Julian},
-  journal={Applied Energy},
-  volume={357},
-  pages={122484},
-  year={2024},
-  doi = {10.1016/j.apenergy.2023.122484},
-  publisher={Elsevier}
-}
 @misc{TimeStruct,
   author       = {Flatberg, Truls and Hellemo, Lars},
   title        = {{TimeStruct.jl: Flexible time structures in optimization modelling}},
