### 2025-02-14 #7578 @trulsf
@vissarion I have now removed the duplicate reference from the paper's bibliography.
