### 2025-03-03 #7578 @crvernon
:wave: @trulsf - I just need you to address the following before I move to accept this for publication:

- [ ] LINES 22-23:  Wording is off a little here.  This should instead read as follows:  "...scenarios.  One example of these types of constraints is in keeping track of a storage inventory over time or incorporating dispatch constraints."
- [ ] LINES 37-38:  "...(e.g., (Loulou et al., 2016), (Tejada-Arango et al. 2024)), ..." should be formatted as "...(e.g., (Loulou et al., 2016; Tejada-Arango et al. 2024))"
- [ ] LINE 89:  "north sea" should be written "North Sea" ... you can maintain the formatting of characters in the bib file using curly brackets around them.
- [ ] LINE 112:  The "n" in "norway" should be capitalized

Let me know when these have been addressed and I will move to accept.  Thanks.
