+++ docs/papers/joss/paper.bib
@@ -1,5 +1,5 @@
 @article{svendsmark2024developing,
-  title={Developing hydrogen energy hubs: The role of H2 prices, wind power and infrastructure investments in Northern Norway},
+  title={Developing hydrogen energy hubs: The role of $\mathrm{H_2}$ prices, wind power and infrastructure investments in {N}orthern {N}orway},
   author={Svendsmark, Erik and Straus, Julian and del Granado, Pedro Crespo},
   journal={Applied Energy},
   volume={376},
@@ -18,7 +18,7 @@
 }
 
 @article{bodal2024hydrogen,
-  title={Hydrogen for harvesting the potential of offshore wind: A North Sea case study},
+  title={Hydrogen for harvesting the potential of offshore wind: A {N}orth {S}ea case study},
   author={B{\o}dal, Espen Flo and Holm, Sigmund Eggen and Subramanian, Avinash and Durakovic, Goran and Pinel, Dimitri and Hellemo, Lars and Ortiz, Miguel Mu{\~n}oz and Knudsen, Brage Rugstad and Straus, Julian},
   journal={Applied Energy},
   volume={357},
